pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wearable_stress_biomarker"
version = "0.1.0"
description = "A production-grade portfolio project for wearable stress detection using WESAD."
authors = [
    {name = "Giulio Matteucci", email = "giulio.matteucci@example.com"},
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "torch>=1.10.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "pyyaml>=6.0",
    "fastapi>=0.70.0",
    "uvicorn>=0.15.0",
    "python-multipart",
    "joblib>=1.1.0",
    "pyarrow>=14.0.0"
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "httpx>=0.24.0",
    "tqdm>=4.65.0",
    "pre-commit>=3.5.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"notebooks/*.ipynb" = ["E402"]
"tests/*" = ["E402"]
